+++ tsdist/lib/types.js
@@ -1,0 +1,3 @@
+/* --------------------------------- ADDONS --------------------------------- */
+"use strict";
+//# sourceMappingURL=types.js.map
